Load global status data separately
- Create new global state manager
- Load *after* login
- Prevents auth popup dialog and failure messages
