+++ src/frontend/src/states/ApiState.tsx
@@ -2,20 +2,15 @@ import { create } from 'zustand';
 import { createJSONStorage, persist } from 'zustand/middleware';
 
 import { api } from '../App';
-import { StatusCodeListInterface } from '../components/render/StatusRenderer';
 import { statusCodeList } from '../defaults/backendMappings';
 import { emptyServerAPI } from '../defaults/defaults';
 import { ApiEndpoints } from '../enums/ApiEndpoints';
-import { ModelType } from '../enums/ModelType';
 import { AuthProps, ServerAPIProps } from './states';
-
-type StatusLookup = Record<ModelType | string, StatusCodeListInterface>;
 
 interface ServerApiStateProps {
   server: ServerAPIProps;
   setServer: (newServer: ServerAPIProps) => void;
   fetchServerApiState: () => void;
-  status?: StatusLookup;
   auth_settings?: AuthProps;
 }
 
@@ -31,19 +26,9 @@ export const useServerApiState = create<ServerApiStateProps>()(
           .then((response) => {
             set({ server: response.data });
           })
-          .catch(() => {});
-        // Fetch status data for rendering labels
-        await api
-          .get(apiUrl(ApiEndpoints.global_status))
-          .then((response) => {
-            const newStatusLookup: StatusLookup = {} as StatusLookup;
-            for (const key in response.data) {
-              newStatusLookup[statusCodeList[key] || key] =
-                response.data[key].values;
-            }
-            set({ status: newStatusLookup });
-          })
-          .catch(() => {});
+          .catch(() => {
+            console.error('Error fetching server info');
+          });
 
         // Fetch login/SSO behaviour
         await api
@@ -53,7 +38,9 @@ export const useServerApiState = create<ServerApiStateProps>()(
           .then((response) => {
             set({ auth_settings: response.data });
           })
-          .catch(() => {});
+          .catch(() => {
+            console.error('Error fetching SSO information');
+          });
       },
       status: undefined
     }),
